refactor: move jsonSchemaToTsConst in separate module
